Extract config for oss storage
... to make it aligned with other storages.
